Авторизация на отборах, комментариях к ним, стеках + языки в отборах (#11)

+++ src/main/java/ru/hits/internship/interview/service/impl/InterviewServiceImpl.java
@@ -5,6 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.hits.internship.common.exceptions.ForbiddenException;
 import ru.hits.internship.common.exceptions.NotFoundException;
 import ru.hits.internship.common.models.pagination.PagedListDto;
 import ru.hits.internship.interview.entity.InterviewEntity;
@@ -14,12 +15,26 @@ import ru.hits.internship.interview.models.InterviewDto;
 import ru.hits.internship.interview.models.UpdateInterviewDto;
 import ru.hits.internship.interview.repository.InterviewRepository;
 import ru.hits.internship.interview.service.InterviewService;
+import ru.hits.internship.language.entity.LanguageEntity;
+import ru.hits.internship.language.repository.LanguageRepository;
 import ru.hits.internship.partner.entity.CompanyPartnerEntity;
 import ru.hits.internship.partner.repository.CompanyPartnerRepository;
 import ru.hits.internship.stack.entity.StackEntity;
 import ru.hits.internship.stack.repository.StackRepository;
+import ru.hits.internship.user.model.common.UserRole;
+import ru.hits.internship.user.model.dto.user.AuthUser;
+import ru.hits.internship.user.model.entity.UserEntity;
+import ru.hits.internship.user.model.entity.role.StudentEntity;
+import ru.hits.internship.user.repository.UserRepository;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
+
+import static ru.hits.internship.interview.service.common.InterviewUtils.getStudentIdIfExists;
+import static ru.hits.internship.interview.service.common.InterviewUtils.isUserAuthor;
+import static ru.hits.internship.user.utils.RoleChecker.isUserHasRole;
 
 @Service
 @RequiredArgsConstructor
@@ -28,35 +43,33 @@ public class InterviewServiceImpl implements InterviewService {
     private final InterviewRepository interviewRepository;
     private final InterviewMapper interviewMapper;
     private final StackRepository stackRepository;
+    private final LanguageRepository languageRepository;
     private final CompanyPartnerRepository companyPartnerRepository;
-    //TODO Прикрутить репозиторий (feature/#3932?)
-    //private final StudentRepository studentRepository;
+    private final UserRepository userRepository;
 
     @Override
     @Transactional
-    public InterviewDto createInterview(UUID studentId, CreateInterviewDto createInterviewDto) {
+    public InterviewDto createInterview(AuthUser user, CreateInterviewDto createInterviewDto) {
         InterviewEntity entity = interviewMapper.map(createInterviewDto);
 
-        UUID stackId = createInterviewDto.getStackId();
-        if (!stackRepository.existsById(stackId)) {
-            throw new NotFoundException("Стек с id %s не найден".formatted(stackId));
-        }
-        StackEntity stack = stackRepository.getReferenceById(stackId);
+        StackEntity stack = getStackReference(createInterviewDto.getStackId());
         entity.setStack(stack);
 
-        UUID companyPartnerId = createInterviewDto.getCompanyPartnerId();
-        if (!companyPartnerRepository.existsById(companyPartnerId)) {
-            throw new NotFoundException("Партнер с id %s не найден".formatted(companyPartnerId));
-        }
-        CompanyPartnerEntity companyPartner = companyPartnerRepository.getReferenceById(companyPartnerId);
+        List<LanguageEntity> languages = getLanguageReferences(createInterviewDto.getLanguageIds());
+        entity.setLanguages(languages);
+
+        CompanyPartnerEntity companyPartner = getCompanyPartnerReference(createInterviewDto.getCompanyPartnerId());
         entity.setCompany(companyPartner);
 
-        //TODO Прикрутить репозиторий (feature/#3932?)
-        //if (!studentRepository.existsById(studentId)) {
-        //    throw new NotFoundException("Студент с id %s не найден".formatted(studentId));
-        //}
-        //StudentEntity student = studentRepository.getReferenceById(studentId);
-        //entity.setStudent(student);
+        UserEntity userEntity = userRepository.findById(user.id())
+                .orElseThrow(() -> new NotFoundException("Пользователь с id %s не найден".formatted(user.id())));
+        StudentEntity student = userEntity.getRoles()
+                .stream()
+                .filter(role -> role.getUserRole() == UserRole.STUDENT)
+                .map(role -> (StudentEntity) role)
+                .findFirst()
+                .orElseThrow(() -> new NotFoundException("Пользователь не имеет роли студента"));
+        entity.setStudent(student);
 
         InterviewEntity savedEntity = interviewRepository.save(entity);
         return interviewMapper.map(savedEntity);
@@ -64,18 +77,21 @@ public class InterviewServiceImpl implements InterviewService {
 
     @Override
     @Transactional
-    public InterviewDto updateInterview(UUID interviewId, UpdateInterviewDto updateInterviewDto) {
+    public InterviewDto updateInterview(AuthUser user, UUID interviewId, UpdateInterviewDto updateInterviewDto) {
         InterviewEntity interview = interviewRepository.findById(interviewId)
                 .orElseThrow(() -> new NotFoundException("Интервью с id %s не найдено".formatted(interviewId)));
 
+        if (!isUserAuthor(user, interview) && !isUserHasRole(user, UserRole.DEAN)) {
+            throw new ForbiddenException();
+        }
+
         interviewMapper.update(interview, updateInterviewDto);
 
-        UUID stackId = updateInterviewDto.getStackId();
-        if (!stackRepository.existsById(stackId)) {
-            throw new NotFoundException("Стек с id %s не найден".formatted(stackId));
-        }
-        StackEntity stack = stackRepository.getReferenceById(stackId);
+        StackEntity stack = getStackReference(updateInterviewDto.getStackId());
         interview.setStack(stack);
+
+        List<LanguageEntity> languages = getLanguageReferences(updateInterviewDto.getLanguageIds());
+        interview.setLanguages(languages);
 
         InterviewEntity savedInterview = interviewRepository.save(interview);
 
@@ -84,26 +100,80 @@ public class InterviewServiceImpl implements InterviewService {
 
     @Override
     @Transactional
-    public void deleteInterview(UUID interviewId) {
+    public void deleteInterview(AuthUser user, UUID interviewId) {
         InterviewEntity interview = interviewRepository.findById(interviewId)
                 .orElseThrow(() -> new NotFoundException("Интервью с id %s не найдено".formatted(interviewId)));
+
+        if (!isUserAuthor(user, interview) && !isUserHasRole(user, UserRole.DEAN)) {
+            throw new ForbiddenException();
+        }
 
         interviewRepository.delete(interview);
     }
 
     @Override
     @Transactional(readOnly = true)
-    public InterviewDto getInterview(UUID interviewId) {
-        return interviewRepository.findById(interviewId)
-                .map(interviewMapper::map)
+    public InterviewDto getInterview(AuthUser user, UUID interviewId) {
+        InterviewEntity interview = interviewRepository.findById(interviewId)
                 .orElseThrow(() -> new NotFoundException("Интервью с id %s не найдено".formatted(interviewId)));
+
+        if (!isUserAuthor(user, interview) && !isUserHasRole(user, UserRole.DEAN) && !isUserHasRole(user, UserRole.CURATOR)) {
+            throw new ForbiddenException();
+        }
+
+        return interviewMapper.map(interview);
     }
 
     @Override
+    public PagedListDto<InterviewDto> getInterviewList(AuthUser user, UUID studentId, Pageable pageable) {
+        if (!isUserHasRole(user, UserRole.DEAN) && !isUserHasRole(user, UserRole.CURATOR)) {
+            throw new ForbiddenException();
+        }
+
+        return getInterviewListByStudentId(studentId, pageable);
+    }
+
+    @Override
+    public PagedListDto<InterviewDto> getInterviewList(AuthUser user, Pageable pageable) {
+        Optional<UUID> studentId = getStudentIdIfExists(user);
+
+        if (studentId.isEmpty()) {
+            throw new NotFoundException("Пользователь не имеет роли студента");
+        }
+
+        return getInterviewListByStudentId(studentId.get(), pageable);
+    }
+
     @Transactional(readOnly = true)
-    public PagedListDto<InterviewDto> getInterviewList(UUID studentId, Pageable pageable) {
+    private PagedListDto<InterviewDto> getInterviewListByStudentId(UUID studentId, Pageable pageable) {
         Page<InterviewEntity> page = interviewRepository.findAllByStudentId(studentId, pageable);
 
         return new PagedListDto<>(page.map(interviewMapper::map));
+    }
+
+    private CompanyPartnerEntity getCompanyPartnerReference(UUID companyPartnerId) {
+        if (!companyPartnerRepository.existsById(companyPartnerId)) {
+            throw new NotFoundException("Партнер с id %s не найден".formatted(companyPartnerId));
+        }
+        return companyPartnerRepository.getReferenceById(companyPartnerId);
+    }
+
+    private StackEntity getStackReference(UUID stackId) {
+        if (!stackRepository.existsById(stackId)) {
+            throw new NotFoundException("Стек с id %s не найден".formatted(stackId));
+        }
+        return stackRepository.getReferenceById(stackId);
+    }
+
+    private List<LanguageEntity> getLanguageReferences(List<UUID> languageIds) {
+        List<LanguageEntity> languages = new ArrayList<>();
+        for (UUID languageId : languageIds) {
+            if (!languageRepository.existsById(languageId)) {
+                throw new NotFoundException("Язык с id %s не найден".formatted(languageId));
+            }
+            LanguageEntity language = languageRepository.getReferenceById(languageId);
+            languages.add(language);
+        }
+        return languages;
     }
 }
